LPD-44398 SF, rename package
