+++ modules/apps/captcha/captcha-impl/src/main/java/com/liferay/captcha/internal/function/captcha/FunctionCaptchaImplFactory.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.captcha.internal.client.extension;
+package com.liferay.captcha.internal.function.captcha;
 
 import com.liferay.osgi.util.configuration.ConfigurationFactoryUtil;
 import com.liferay.petra.string.StringBundler;
@@ -57,7 +57,7 @@ public class FunctionCaptchaImplFactory {
 	}
 
 	@Reference(
-		target = "(component.factory=com.liferay.captcha.internal.client.extension.FunctionCaptchaImpl)"
+		target = "(component.factory=com.liferay.captcha.internal.function.captcha.FunctionCaptchaImpl)"
 	)
 	private ComponentFactory<FunctionCaptchaImpl> _componentFactory;
